backend/requirements.txt
django==5.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gitpython==3.1.41
pillow==10.2.0
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-compressor==4.4

# API & Utils
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.1